[TASK] Move html into a separate file
