+++ src/pages-sync/trackingMode/VideoStrategy.ts
@@ -2,6 +2,7 @@ import { isIframeUrl } from '../../utils/manifest';
 import { hasMissingPermissions } from '../../utils/customDomains';
 import { PlayerSingleton } from '../../utils/player';
 import { ProgressElement, TrackingModeInterface } from './TrackingModeInterface';
+import { videoStrategyErrorElement } from '../messageElements';
 
 export class VideoStrategy implements TrackingModeInterface {
   trackingResolve: (() => void) | null = null;
@@ -34,76 +35,22 @@ export class VideoStrategy implements TrackingModeInterface {
     let discordTimeout;
 
     this.playerFoundTimeout = setTimeout(
-      async () => {
+      (async () => {
         if (this.errorListener) {
-          const errorDiv = document.createElement('div');
-          errorDiv.className = 'player-error';
-          errorDiv.style.cssText = `
-          display: block;
-          padding: 5px;
-          padding-top: 15px;
-          outline: 1px solid #e13f7b;
-        `;
-
-          const linkContainer = document.createElement('div');
-          linkContainer.style.cssText = `
-          display: flex;
-          justify-content: space-evenly;
-        `;
-
-          if (await hasMissingPermissions()) {
-            const missingPermissionsSpan = document.createElement('span');
-            missingPermissionsSpan.className = 'player-error-missing-permissions';
-            missingPermissionsSpan.style.cssText = `
-          padding-top: 10px;
-        `;
-            missingPermissionsSpan.textContent = api.storage.lang(
-              'settings_custom_domains_missing_permissions_header',
-            );
-
-            errorDiv.appendChild(missingPermissionsSpan);
-
-            const addLink = document.createElement('a');
-            addLink.className = 'player-error-missing-permissions';
-            addLink.href = 'https://malsync.moe/pwa/#/settings/customDomains';
-            addLink.style.cssText = `
-          margin: 10px;
-          border-bottom: 2px solid #e13f7b;
-        `;
-            addLink.textContent = api.storage.lang('Add');
-
-            linkContainer.appendChild(addLink);
-          } else {
-            const defaultSpan = document.createElement('span');
-            defaultSpan.className = 'player-error-default';
-            defaultSpan.textContent = api.storage.lang('syncPage_flash_player_error');
-
-            errorDiv.appendChild(defaultSpan);
-          }
-
-          const helpLink = document.createElement('a');
-          helpLink.href = 'https://discord.com/invite/cTH4yaw';
-          helpLink.style.cssText = `
-          display: block;
-          margin: 10px;
-        `;
-          helpLink.textContent = 'Help';
-
-          linkContainer.appendChild(helpLink);
-
-          errorDiv.appendChild(linkContainer);
-
+          const errorDiv = videoStrategyErrorElement(await hasMissingPermissions());
           this.errorListener(errorDiv);
 
           const iframes = $('iframe')
             .toArray()
-            .map(el => utils.absoluteLink($(el).attr('src'), window.location.origin))
+            .map(
+              el => utils.absoluteLink($(el).attr('src'), window.location.origin) as string | null,
+            )
             .filter(el => el)
-            .filter(el => !isIframeUrl(el));
+            .filter(el => !isIframeUrl(el!));
 
           con.log('No Player found', iframes);
         }
-      },
+      }) as () => void,
       5 * 60 * 1000,
     );
     PlayerSingleton.getInstance()
